Upstream version 0.5.2

# gpg: Signature made Tue Jan 13 23:16:55 2015 EST using RSA key ID B0B3C74B
# gpg: Good signature from "Team Vuo <info@vuo.org>"

+++ type/VuoPoint2d.h
@@ -91,8 +91,8 @@ static inline float VuoPoint2d_squaredMagnitude(VuoPoint2d a)
 /**
  * @c a / @c b
  */
-static inline VuoPoint2d VuoPoint2d_divide(VuoPoint2d a, double b) __attribute__((const));
-static inline VuoPoint2d VuoPoint2d_divide(VuoPoint2d a, double b)
+static inline VuoPoint2d VuoPoint2d_divide(VuoPoint2d a, float b) __attribute__((const));
+static inline VuoPoint2d VuoPoint2d_divide(VuoPoint2d a, float b)
 {
 	VuoPoint2d p =
 	{
@@ -130,8 +130,8 @@ static inline VuoPoint2d VuoPoint2d_normalize(VuoPoint2d a)
 /**
  * @c a * @c b
  */
-static inline VuoPoint2d VuoPoint2d_multiply(VuoPoint2d a, double b) __attribute__((const));
-static inline VuoPoint2d VuoPoint2d_multiply(VuoPoint2d a, double b)
+static inline VuoPoint2d VuoPoint2d_multiply(VuoPoint2d a, float b) __attribute__((const));
+static inline VuoPoint2d VuoPoint2d_multiply(VuoPoint2d a, float b)
 {
 	VuoPoint2d p =
 	{
